+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Some context: this release heavily relies on where clauses on contextually gener
 ### Deprecated
 
 * Several cell/header/footer/supplementary view registration methods have been deprecated to unify registration logic. Please use `register(_:handler:mapping:)`, `registerHeader(_:handler:mapping:)`, `registerFooter(_:handler:mapping:)` as a replacements for all of those methods. For more information on those changes, please read [migration guide LINK MISSING](link to migration guide).
+* All non-deprecated registration methods now have an additional `handler` closure, that allows to configure cells/headers/footers that are dequeued from UITableView. This is a direct replacement for `configure(_:_:`, `configureHeader(_:_:)`, `configureFooter(_:_:)` , that are all now deprecated. Please note, that handler closure is called before `DTModelTransfer.update(with:)` method.
 
 ## [7.2.0](https://github.com/DenTelezhkin/DTTableViewManager/releases/tag/7.2.0)